refactor(combat): split probabilistic resolver types and casualties (#4545)
Extract ProbabilisticRoundResult, ProbabilisticEngagementOutcome, and round
constants into sibling libraries; move Poisson sampling and weighted casualty
selection into a dedicated helper. Keeps resolveEngagementProbabilistic as the
barrel-exported entry with stable import paths (Slice A).
